third_party/blink/web_tests/platform/generic/svg/custom/path-bbox-update-expected.txt

PASS: bounding box is (80, 80, 100, 100)
